feat(domain): sparse, unbounded TileLayer
Replace the fixed-size grid with a sparse board. TileLayer now stores only
painted cells in an unordered_map keyed by signed (x,y), so the canvas is
unbounded in every direction and empty cells cost nothing. LayerStack drops its
shared width/height; both gain GetBounds() for framing and export.

Coordinates are rationalised: a tile at (cx,cy) lives at world centre
(cx*tileSize, cy*tileSize), origin (0,0) is a real anchor, and the old
'+1 border / 1-based' scheme is gone. Painting goes through SetTile via the
paint/erase/fill commands; flood fill is bounded to painted content + the click
cell. Camera framing reflows on origin/content-bounds; export sizes to the
content bbox. The fixed-size project UI (resize, dimension readouts) is removed.

Serialization stores a sparse [{x,y,...}] tile list, with a legacy-dense
loader so existing .json projects still open.

+++ Tiles/src/Tiles/Core/Session/EditingState.cpp
@@ -31,7 +31,7 @@ namespace Tiles
         }
     }
 
-    std::unique_ptr<Command> EditingState::BuildModeCommand(size_t layerIndex, size_t x, size_t y, const Tile& tile) const
+    std::unique_ptr<Command> EditingState::BuildModeCommand(size_t layerIndex, int x, int y, const Tile& tile) const
     {
         switch (m_PaintingMode)
         {
@@ -46,12 +46,12 @@ namespace Tiles
         }
     }
 
-    std::unique_ptr<Command> EditingState::BuildEraseCommand(size_t layerIndex, size_t x, size_t y) const
+    std::unique_ptr<Command> EditingState::BuildEraseCommand(size_t layerIndex, int x, int y) const
     {
         return std::make_unique<TileEraseCommand>(x, y, layerIndex);
     }
 
-    std::unique_ptr<Command> EditingState::BuildFillCommand(size_t layerIndex, size_t x, size_t y, const Tile& tile) const
+    std::unique_ptr<Command> EditingState::BuildFillCommand(size_t layerIndex, int x, int y, const Tile& tile) const
     {
         return std::make_unique<LayerFillCommand>(x, y, layerIndex, tile);
     }
