+++ Tiles/src/Tiles/Core/Session/EditingState.h
@@ -45,9 +45,9 @@ namespace Tiles
         void ValidateWorkingLayer(const LayerStack& layerStack);
 
         /// Builds the command for the current painting mode; null for None.
-        std::unique_ptr<Command> BuildModeCommand(size_t layerIndex, size_t x, size_t y, const Tile& tile) const;
-        std::unique_ptr<Command> BuildEraseCommand(size_t layerIndex, size_t x, size_t y) const;
-        std::unique_ptr<Command> BuildFillCommand(size_t layerIndex, size_t x, size_t y, const Tile& tile) const;
+        std::unique_ptr<Command> BuildModeCommand(size_t layerIndex, int x, int y, const Tile& tile) const;
+        std::unique_ptr<Command> BuildEraseCommand(size_t layerIndex, int x, int y) const;
+        std::unique_ptr<Command> BuildFillCommand(size_t layerIndex, int x, int y, const Tile& tile) const;
 
     private:
         size_t m_WorkingLayer = 0;
